♻️ move types and fix tests
