+++ dashboard-app/src/components/StackedBarChart/types.ts
@@ -1,0 +1,7 @@
+export interface LegendDatum {
+  id: number;
+  name: string;
+  active: boolean;
+}
+
+export type LegendData = LegendDatum[];
