+++ dashboard-app/src/components/StackedBarChart/utils/__tests__/index.test.ts
@@ -15,24 +15,24 @@ describe('Helpers', () => {
   describe(':: isEveryDatumActive', async () => {
     test('Success :: return true if all data is active', async () => {
       const activeData = [
-        { active: true, key: 'Aku Aku' },
-        { active: true, key: 'Crash Bandicoot' },
+        { active: true, name: 'Aku Aku', id: 7 },
+        { active: true, name: 'Crash Bandicoot', id: 3 },
       ];
       const expected = isEveryDatumActive(activeData);
       expect(expected).toBe(true);
     });
     test('Success :: return false if one datum is inactive', async () => {
       const activeData = [
-        { active: true, key: 'Aku Aku' },
-        { active: false, key: 'Crash Bandicoot' },
+        { active: true, name: 'Aku Aku', id: 6 },
+        { active: false, name: 'Crash Bandicoot', id: 3 },
       ];
       const expected = isEveryDatumActive(activeData);
       expect(expected).toBe(false);
     });
     test('Success :: return false if all data is inactive', async () => {
       const activeData = [
-        { active: false, key: 'Aku Aku' },
-        { active: false, key: 'Crash Bandicoot' },
+        { active: false, name: 'Aku Aku', id: 2 },
+        { active: false, name: 'Crash Bandicoot', id: 5 },
       ];
       const expected = isEveryDatumActive(activeData);
       expect(expected).toBe(false);
@@ -42,24 +42,24 @@ describe('Helpers', () => {
   describe(':: isEveryDatumInactive', async () => {
     test('Success :: return true if all data is inactive', async () => {
       const activeData = [
-        { active: false, key: 'Aku Aku' },
-        { active: false, key: 'Crash Bandicoot' },
+        { active: false, name: 'Aku Aku', id: 3 },
+        { active: false, name: 'Crash Bandicoot', id: 2 },
       ];
       const expected = isEveryDatumInactive(activeData);
       expect(expected).toBe(true);
     });
     test('Success :: return false if one datum is active', async () => {
       const activeData = [
-        { active: true, key: 'Aku Aku' },
-        { active: false, key: 'Crash Bandicoot' },
+        { active: true, name: 'Aku Aku', id: 2 },
+        { active: false, name: 'Crash Bandicoot', id: 6 },
       ];
       const expected = isEveryDatumInactive(activeData);
       expect(expected).toBe(false);
     });
     test('Success :: return false if all data is active', async () => {
       const activeData = [
-        { active: true, key: 'Aku Aku' },
-        { active: true, key: 'Crash Bandicoot' },
+        { active: true, name: 'Aku Aku', id: 2 },
+        { active: true, name: 'Crash Bandicoot', id: 6 },
       ];
       const expected = isEveryDatumInactive(activeData);
       expect(expected).toBe(false);
@@ -69,35 +69,35 @@ describe('Helpers', () => {
   describe(':: flipActiveOfAll', async () => {
     test('Success :: sets all to false if all are true', async () => {
       const activeData = [
-        { active: true, key: 'Aku Aku' },
-        { active: true, key: 'Crash Bandicoot' },
+        { active: true, name: 'Aku Aku', id: 2 },
+        { active: true, name: 'Crash Bandicoot', id: 6 },
       ];
       const expected = flipActiveOfAll(activeData);
       expect(expected).toEqual([
-        { active: false, key: 'Aku Aku' },
-        { active: false, key: 'Crash Bandicoot' },
+        { active: false, name: 'Aku Aku', id: 2 },
+        { active: false, name: 'Crash Bandicoot', id: 6 },
       ]);
     });
     test('Success :: sets all to false if one is false', async () => {
       const activeData = [
-        { active: false, key: 'Aku Aku' },
-        { active: true, key: 'Crash Bandicoot' },
+        { active: false, name: 'Aku Aku', id: 2 },
+        { active: true, name: 'Crash Bandicoot', id: 6 },
       ];
       const expected = flipActiveOfAll(activeData);
       expect(expected).toEqual([
-        { active: false, key: 'Aku Aku' },
-        { active: false, key: 'Crash Bandicoot' },
+        { active: false, name: 'Aku Aku', id: 2 },
+        { active: false, name: 'Crash Bandicoot', id: 6 },
       ]);
     });
     test('Success :: sets all to true if all are false', async () => {
       const activeData = [
-        { active: false, key: 'Aku Aku' },
-        { active: false, key: 'Crash Bandicoot' },
+        { active: false, name: 'Aku Aku', id: 2 },
+        { active: false, name: 'Crash Bandicoot', id: 6 },
       ];
       const expected = flipActiveOfAll(activeData);
       expect(expected).toEqual([
-        { active: true, key: 'Aku Aku' },
-        { active: true, key: 'Crash Bandicoot' },
+        { active: true, name: 'Aku Aku', id: 2 },
+        { active: true, name: 'Crash Bandicoot', id: 6 },
       ]);
     });
   });
